NavigationRouteCallback, NavigationWalkingOptions, NavigationRoute: move to `base` module
